Close five audited gaps: launchers, correctness, replay, licensing, serving
SFT, DPO and GRPO had no way to reach a GPU on this cluster, which is why they
were never proven. All three now have sbatch launchers, and the SFT path is
proven on real hardware: job 27049 ran configs/sft_14b_full.json verbatim on 8
ranks with n_rows=71510 and dropped_overlong=66 -- exactly what the readiness
plan predicts for max_seq_length 17408 -- on a 1,677-step horizon with a 5h40m
trainer ETA, then drained cleanly on cancel. A shared resolver does the
model_id/output_dir override so one command is the whole stage and a requeued
child re-derives the same config, and the previous stage's consolidated weights
are checked before eight ranks each load 55 GiB.

The replay cache is reachable for the first time. kore/env/preflight_identity.py
validates GPUs out of process and mints one content-addressed identity per card;
a duplicate candidate now costs 2.8 ms instead of ~6 s and replay_hits moves off
a permanent zero. Keeping the producer outside KoreEnv is what makes it
non-circular: an identity minted at evaluation time, by the same process, from
the same state it authorizes, could never fail. Wiring it also exposed that
collect_amd_gpu_devices could never resolve, because the HIP half of its
BDF join only came from an env var nothing produced.

Two real campaign bugs surfaced from writing tests for the five stages that had
none. The DPO >=8% hard-negative floor was computed, logged, and then ignored --
a shipped 14B run trained on 4.1% and data/full14b/events.jsonl records it. And
sequential datagen published with bare write_jsonl, so its own resume check
could never return True and the production build reader rejected its records
outright, meaning --datagen-workers 1 could not build at all. Both now fail
closed. The held-out leak tests assert on the emitted corpus via a canary
string, not on the filter.

The end-to-end serving gate produced this repository's first evaluation
artifacts, against a real SGLang server on gfx950. The null control -- the same
server measured twice -- is REJECTED, so the gate is not a rubber stamp.

Suite: 7,763 tests green including the release gates.

+++ kore/tasks/breadth/ssm_ext.py
@@ -1296,8 +1296,13 @@ def _ssm_lru_kernel(x_ptr, lr_ptr, li_ptr, y_ptr, D, L, srow, sl):
 
 def ssm_lru(x: torch.Tensor, nu: torch.Tensor, theta: torch.Tensor) -> torch.Tensor:
     B, D, L, _ = x.shape
-    lr = (nu * torch.cos(theta)).contiguous()
-    li = (nu * torch.sin(theta)).contiguous()
+    # lam = nu*exp(i*theta) is the recurrence COEFFICIENT, so it must be built in
+    # fp32 like the oracle does, not in the task dtype.  |lam| is biased close to
+    # 1, and a perturbation of lam is amplified by the effective memory
+    # |lam|/(1-|lam|) over L steps: one bf16 rounding of lam (2**-9) turns into
+    # percent-level error in y, which no amount of fp32 state can undo.
+    lr = (nu.float() * torch.cos(theta.float())).contiguous()
+    li = (nu.float() * torch.sin(theta.float())).contiguous()
     xf = x.contiguous().reshape(B * D, L, 2)
     y = torch.empty_like(xf)
     _ssm_lru_kernel[(B * D,)](xf, lr, li, y, D, L, xf.stride(0), xf.stride(1), num_warps=1)
